Added a "generic command" function to the PTZ proxy.  The idea is this is an easy way to send camera config commands to the sony EVI D30/100 without making a bunch of abstract, general camera configs.  We can worry about that when we have multiple PTZs...  The generic here refers to the nature of the command being sent , but it's a bit misleading since this is very sony-specific.  maybe i'll fix th at later.  There's also some code in there about a second movement mode, which i s an attempt to make more smooth movements, but it's not quite working.  It's of f by default.
Changed the MCom driver to add a "capacity" to the buffers of a given type/channel.  The capacity must be < MCOM_N_BUFS (currently 10).  By setting it to 1, we can guarantee to always get the newest data.  Older messages are just written over.  Also changed the command member in player_mcom_config_t to uint8_t instead of uint16_t.


git-svn-id: svn://svn.code.sf.net/p/playerstage/svn/code/player/trunk@1674 9fbd2ccc-adb5-4725-87dd-eba157860cdb